requirements.txt
django==3
django-filter
djangorestframework
django-cors-headers
drf-yasg
drf-yasg[validation]
Pillow
